FIX: Makes ExecCodeProvider::selectArchForFAT match the design spec
The function didn't quite match what was intended for it. A TargetArch t
was passed -- as a value -- and then overwritten inside the code. It was
never checked to make any decisions.

From discussions spawned in slack, I could figure that we wanted to use
t to select one of the available architechtures in the FAT header.

Conflicts:
	libs/getExecSections/getExec.cpp